Update core library handlebars to version 3.0.3.
